Bivariate Functions

AsymmetricRisk.cokurtosis — Method

cokurtosis(x,y;kind::Symbol=:asymmetric)

Computes the cokurtosis between vectors x and y. The default is asymmetric cokurtosis.

kind=:asymmetric: $\frac{E[(x - E[x])^3 (y-E[y])]}{\sigma_x^3 \sigma_y}$

kind=:symmetric: $\frac{E[(x - E[x])^2 (y-E[y])^2]}{\sigma_x^2 \sigma_y^2}$

AsymmetricRisk.coskewness — Method

coskewness(x,y)

Computes the coskewness between vectors x and y. The vector x receives the powers in the expression.

$\frac{E[(x - E[x])^2 (y-E[y])]}{\sigma_x^2 \sigma_y}$

AsymmetricRisk.downsidebeta — Method

downsidebeta(x,y;kind=:both)

Computes the downside beta of y on x as defined in Ang & Chen.

kind=:onx allows to condition only on x.

AsymmetricRisk.exceedancecor — Method

exceedancecor(x,y,c;kind=:both,side=:lower)

Computes exceedance correlations between x and y, using exceedance threshold c. Keyword kind refers to which type of conditioning: :both conditions on x and y simultaneously, while :onx conditions only on x. Keyword side refers to look below (:lower) or above (:upper) the threshold c. Inputs x and y are standardized before conditioning, so c should be in standard deviation units.

AsymmetricRisk.quantiledep — Method

quantiledep(x,y,q)

Quantile Dependence at the q-th quantile of vector-like data x and y.

Let $q_x$ and $q_y$ be the q-th quantiles of x and y.

Returns $E[xy | x \leq q_x, y \leq q_y] / q$ if q<=0.5 and $(1 -2q + E[xy | x \leq q_x, y \leq q_y]) / (1-q)$ otherwise.

AsymmetricRisk.upsidebeta — Method

upsidebeta(x,y;kind=:both)

Computes the upside beta of y on x as defined in Ang & Chen.

kind=:onx allows to condition only on x.